uefi: extract time code to new module
